requirements.txt
Django==3.2.6
djangorestframework==3.12.4
mysqlclient==2.1.1
requests==2.26.0
python-dateutil==2.8.2
pytz==2021.1
